search: rename lib synchronous_product to search
