+++ src/CMakeLists.txt
@@ -2,6 +2,6 @@ add_subdirectory(automata)
 add_subdirectory(mtl)
 add_subdirectory(mtl_ata_translation)
 add_subdirectory(utilities)
-add_subdirectory(synchronous_product)
+add_subdirectory(search)
 add_subdirectory(visualization)
 add_subdirectory(app)
