+++ src/app/app.cpp
@@ -30,10 +30,10 @@
 #include "mtl/mtl.pb.h"
 #include "mtl/mtl_proto.h"
 #include "mtl_ata_translation/translator.h"
-#include "synchronous_product/create_controller.h"
-#include "synchronous_product/heuristics.h"
-#include "synchronous_product/search.h"
-#include "synchronous_product/search_tree.h"
+#include "search/create_controller.h"
+#include "search/heuristics.h"
+#include "search/search.h"
+#include "search/search_tree.h"
 #include "visualization/ta_to_graphviz.h"
 #include "visualization/tree_to_graphviz.h"
