+++ src/automata/include/automata/automata.h
@@ -25,6 +25,7 @@
 #include <functional>
 #include <iostream>
 #include <map>
+#include <optional>
 #include <set>
 #include <stdexcept>
 #include <string>
