+++ test/test_synchronous_product.cpp
@@ -24,10 +24,10 @@
 #include "automata/ta_regions.h"
 #include "mtl/MTLFormula.h"
 #include "mtl_ata_translation/translator.h"
-#include "synchronous_product/canonical_word.h"
-#include "synchronous_product/operators.h"
-#include "synchronous_product/reg_a.h"
-#include "synchronous_product/synchronous_product.h"
+#include "search/canonical_word.h"
+#include "search/operators.h"
+#include "search/reg_a.h"
+#include "search/synchronous_product.h"
 #include "utilities/Interval.h"
 #include "utilities/numbers.h"
